Type mappings and refactoring
- Added type_mappings to producer properties, to correspond alias "event" in the consumer configuration
- Packages renamed in order to give more readability
- Added 1 more deep level (added producer) into application properties. Also modified META_INF file.
