+++ src/main/java/cl/atromilen/spring/kafka/producer/SpringKafkaProducerApplication.java
@@ -1,4 +1,4 @@
-package cl.atromilen.springkafkaproducer;
+package cl.atromilen.spring.kafka.producer;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
