+++ src/main/java/cl/atromilen/spring/kafka/producer/config/ConfigProperties.java
@@ -1,11 +1,11 @@
-package cl.atromilen.springkafkaproducer.config;
+package cl.atromilen.spring.kafka.producer.config;
 
 import lombok.Data;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.stereotype.Component;
 
 @Component
-@ConfigurationProperties(prefix = "spring.kafka")
+@ConfigurationProperties(prefix = "spring.kafka.producer")
 @Data
 public class ConfigProperties {
     private String bootstrapServers;
