+++ src/main/java/cl/atromilen/spring/kafka/producer/config/KafkaTopicConfiguration.java
@@ -1,4 +1,4 @@
-package cl.atromilen.springkafkaproducer.config;
+package cl.atromilen.spring.kafka.producer.config;
 
 import lombok.RequiredArgsConstructor;
 import org.apache.kafka.clients.admin.AdminClientConfig;
