+++ src/main/java/cl/atromilen/spring/kafka/producer/controller/MessageController.java
@@ -1,7 +1,7 @@
-package cl.atromilen.springkafkaproducer.controller;
+package cl.atromilen.spring.kafka.producer.controller;
 
-import cl.atromilen.springkafkaproducer.event.MessageForEmailing;
-import cl.atromilen.springkafkaproducer.service.KafkaProducerService;
+import cl.atromilen.spring.kafka.producer.event.MessageForEmailing;
+import cl.atromilen.spring.kafka.producer.service.KafkaProducerService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
