+++ src/main/resources/application-local.yml
@@ -1,8 +1,9 @@
 spring:
     kafka:
-      bootstrap-servers: localhost:29092
-      topic-name: messages-for-emailing
-      topic-partitions: 1
-      topic-replicas: 1
+      producer:
+        bootstrap-servers: localhost:29092
+        topic-name: messages-for-emailing
+        topic-partitions: 1
+        topic-replicas: 1
 server:
   port: 8081
