+++ src/main/resources/application-test.yml
@@ -1,6 +1,7 @@
 spring:
     kafka:
-      bootstrap-servers: localhost:29092
-      topic-name: embedded-test-topic
-      topic-partitions: 1
-      topic-replicas: 1
+      producer:
+        bootstrap-servers: localhost:29092
+        topic-name: embedded-test-topic
+        topic-partitions: 1
+        topic-replicas: 1
