+++ src/test/java/cl/atromilen/spring/kafka/producer/SpringKafkaProducerApplicationTests.java
@@ -1,4 +1,4 @@
-package cl.atromilen.springkafkaproducer;
+package cl.atromilen.spring.kafka.producer;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
